reorganising tests a bit
